Use a single attribute to split gift and compute cartRules
