+++ src/Core/Domain/Cart/QueryResult/CartForOrderCreation.php
@@ -26,16 +26,16 @@
 
 namespace PrestaShop\PrestaShop\Core\Domain\Cart\QueryResult;
 
-use PrestaShop\PrestaShop\Core\Domain\Cart\QueryResult\CartInformation\CartAddress;
-use PrestaShop\PrestaShop\Core\Domain\Cart\QueryResult\CartInformation\CartProduct;
-use PrestaShop\PrestaShop\Core\Domain\Cart\QueryResult\CartInformation\CartRule;
-use PrestaShop\PrestaShop\Core\Domain\Cart\QueryResult\CartInformation\CartShipping;
-use PrestaShop\PrestaShop\Core\Domain\Cart\QueryResult\CartInformation\CartSummary;
+use PrestaShop\PrestaShop\Core\Domain\Cart\QueryResult\CartForOrderCreation\CartAddress;
+use PrestaShop\PrestaShop\Core\Domain\Cart\QueryResult\CartForOrderCreation\CartProduct;
+use PrestaShop\PrestaShop\Core\Domain\Cart\QueryResult\CartForOrderCreation\CartRule;
+use PrestaShop\PrestaShop\Core\Domain\Cart\QueryResult\CartForOrderCreation\CartShipping;
+use PrestaShop\PrestaShop\Core\Domain\Cart\QueryResult\CartForOrderCreation\CartSummary;
 
 /**
  * Holds cart information data
  */
-class CartInformation
+class CartForOrderCreation
 {
     /**
      * @var int
