+++ src/Core/Domain/Cart/QueryResult/CartForOrderCreation/CartShipping.php
@@ -24,7 +24,7 @@
  * @license   https://opensource.org/licenses/OSL-3.0 Open Software License (OSL 3.0)
  */
 
-namespace PrestaShop\PrestaShop\Core\Domain\Cart\QueryResult\CartInformation;
+namespace PrestaShop\PrestaShop\Core\Domain\Cart\QueryResult\CartForOrderCreation;
 
 use PrestaShop\PrestaShop\Core\Domain\Cart\QueryResult\CartDeliveryOption;
