+++ src/PrestaShopBundle/Controller/Admin/Sell/Order/OrderController.php
@@ -1167,7 +1167,10 @@ class OrderController extends FrameworkBundleAdminController
         $cartId = $this->getCommandBus()->handle(new DuplicateOrderCartCommand($orderId))->getValue();
 
         return $this->json(
-            $this->getQueryBus()->handle(new GetCartForOrderCreation($cartId))
+            $this->getQueryBus()->handle(
+                (new GetCartForOrderCreation($cartId))
+                ->setHideDiscounts(true)
+            )
         );
     }
